+++ src/main/java/racingGame/domain/Number.java
@@ -2,7 +2,9 @@ package racingGame.domain;
 
 public class Number {
     private static final String NEGATIVE_ERROR_MESSAGE = "음수가 들어왔습니다";
+
     private int value;
+
     public Number(int value){
         negativeChk(value);
         this.value = value;
