+++ src/main/java/racingGame/view/InputView.java
@@ -1,6 +1,6 @@
 package racingGame.view;
 
-import racingGame.Number;
+import racingGame.domain.Number;
 
 import java.util.Scanner;
 
@@ -10,14 +10,18 @@ public class InputView {
     private static final String TRY_COUNT_MESSAGE = "시도할 회수는 몇 회 인가요?";
     private static final Scanner sc = new Scanner(System.in);
 
+    private InputView(){
+
+    }
+
     public static int carCount() {
         System.out.println(CAR_COUNT_MESSAGE);
-        return new Number(sc.nextInt()).getValue();
+        return sc.nextInt();
     }
 
     public static int tryCount(){
         System.out.println(TRY_COUNT_MESSAGE);
-        return new Number(sc.nextInt()).getValue();
+        return sc.nextInt();
     }
 
 }
